some fixes and cleanup Patches applied:
 * devel@gnuage.org/mplayer--ui--0--patch-23
   fix warnings

 * devel@gnuage.org/mplayer--ui--0--patch-24
   switched mpui_orientation from #define to enum

 * devel@gnuage.org/mplayer--ui--0--patch-25
   merge from devel@geexbox.org--2004/mplayer--ui--0

+++ libmpui/mpui_struct.c
@@ -133,9 +133,9 @@ mpui_str_get_size (mpui_str_t *str)
   while (*s)
     {
       render_one_glyph (font, *s);
-      f = font->font[*s];
+      f = font->font[(int)*s];
 
-      w += font->width[*s] + font->charspace;
+      w += font->width[(int)*s] + font->charspace;
       if (f >= 0 && font->pic_a[f]->h > h)
         h = font->pic_a[f]->h;
